+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Shopify
 
+## Unreleased
+
+- Fixed a PHP error that occurred when saving invalid settings. ([#39](https://github.com/craftcms/shopify/pull/39),[#40](https://github.com/craftcms/shopify/pull/40))
+
 ## 3.0.0.1 - 2022-11-08
 
 - Fixed a namespacing bug.