+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-- Added `craft\shopify\elements\Product::getDefaultVariant()`.
+- Fixed a PHP error that occurred when saving invalid settings. ([#39](https://github.com/craftcms/shopify/pull/39), [#40](https://github.com/craftcms/shopify/pull/40))
 - Added `craft\shopify\elements\Product::getCheapestVariant()`.
-- Fixed a PHP error that occurred when saving invalid settings. ([#39](https://github.com/craftcms/shopify/pull/39),[#40](https://github.com/craftcms/shopify/pull/40))
+- Added `craft\shopify\elements\Product::getDefaultVariant()`.
 
 ## 3.0.0.1 - 2022-11-08