+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Shopify
 
-## Unreleased
+## 3.0.1 - 2022-11-16
 
 - Fixed a PHP error that occurred when saving invalid settings. ([#39](https://github.com/craftcms/shopify/pull/39), [#40](https://github.com/craftcms/shopify/pull/40))
 - Added `craft\shopify\elements\Product::getCheapestVariant()`.